add send mail from google and we took error:ForbiddenError: invalid csrf token

+++ public/app.js
@@ -48,39 +48,44 @@ if (document.querySelectorAll('.tabs')) {
 }
 
 const CARD_HTML = document.querySelector('#card')
-CARD_HTML.addEventListener('click', (e) => {
-    e.preventDefault()
 
-    if (e.target.classList.contains('js-remove')) {
-        const id = e.target.dataset.id
-        const csrf = e.target.dataset.csrf
-        fetch(`/card/remove/${id}`, {
-            method: 'delete',
-            headers: {
-                'X-XSRF-TOKEN': csrf
-            }
-        }).then(response => response.json())
-            .then(result => {
-                // console.log(e.target)
-                if (result.courses.length) {
-                    const html = result.courses.map(c => {
-                        return `
-                    <tr>
-                    <td>${c.title}</td>
-                    <td id="count">${c.count}</td>
-                    <td class="price total-table__price">${toCurrency(c.price)}</td>
-                    <td>
-                        <button class="btn-floating btn-small waves-effect waves-light red js-remove js-remove material-icons" data-id="${c.id}" data-csrf="${csrf}">remove</button>
-                    </td>
-                    </tr>
-                    `
-                    }).join('')
-                    CARD_HTML.querySelector('tbody').innerHTML = html
-                    document.querySelector('#total-price').textContent = toCurrency(result.total)
-                } else {
-                    CARD_HTML.innerHTML = `<p>Корзина пуста</p>`
+if (CARD_HTML) {
+    CARD_HTML.addEventListener('click', (e) => {
+        e.preventDefault()
+
+        if (e.target.classList.contains('js-remove')) {
+            const id = e.target.dataset.id
+            const csrf = e.target.dataset.csrf
+            fetch(`/card/remove/${id}`, {
+                method: 'delete',
+                headers: {
+                    'X-XSRF-TOKEN': csrf
                 }
-            })
-    }
-})
+            }).then(response => response.json())
+                .then(result => {
+                    // console.log(e.target)
+                    if (result.courses.length) {
+                        const html = result.courses.map(c => {
+                            return `
+                        <tr>
+                        <td>${c.title}</td>
+                        <td id="count">${c.count}</td>
+                        <td class="price total-table__price">${toCurrency(c.price)}</td>
+                        <td>
+                            <button class="btn-floating btn-small waves-effect waves-light red js-remove js-remove material-icons" data-id="${c.id}" data-csrf="${csrf}">remove</button>
+                        </td>
+                        </tr>
+                        `
+                        }).join('')
+                        CARD_HTML.querySelector('tbody').innerHTML = html
+                        document.querySelector('#total-price').textContent = toCurrency(result.total)
+                    } else {
+                        CARD_HTML.innerHTML = `<p>Корзина пуста</p>`
+                    }
+                })
+        }
+    })
+}
+
+
